+++ src/color/color.test.ts
@@ -55,4 +55,12 @@ describe('Color regex and conversion tests', () => {
     expect(color1.isUnrecognised).toBe(true);
     expect(color1.getSource().toString()).toBe([0, 0, 0, 1].toString());
   });
+
+  it('Create colors through keywords', () => {
+    const color1 = new Color('red');
+    const color2 = new Color('RED');
+    const red = 'rgb(255,0,0)';
+    expect(color1.toRgb()).toBe(red);
+    expect(color2.toRgb()).toBe(red);
+  });
 });