+++ src/filters/ColorMatrix.ts
@@ -7,7 +7,7 @@ import type {
 import { classRegistry } from '../ClassRegistry';
 import { fragmentSource } from './shaders/colorMatrix';
 
-type ColorMatrixOwnProps = {
+export type ColorMatrixOwnProps = {
   matrix: TMatColorMatrix;
   colorsOnly: boolean;
 };
@@ -36,7 +36,8 @@ export const colorMatrixDefaultValues: ColorMatrixOwnProps = {
 export class ColorMatrix<
   Name extends string = 'ColorMatrix',
   OwnProps extends object = ColorMatrixOwnProps,
-> extends BaseFilter<Name, OwnProps> {
+  SerializedProps extends object = ColorMatrixOwnProps,
+> extends BaseFilter<Name, OwnProps, SerializedProps> {
   /**
    * Colormatrix for pixels.
    * array of 20 floats. Numbers in positions 4, 9, 14, 19 loose meaning
@@ -81,16 +82,15 @@ export class ColorMatrix<
       const r = data[i];
       const g = data[i + 1];
       const b = data[i + 2];
-      if (colorsOnly) {
-        data[i] = r * m[0] + g * m[1] + b * m[2] + m[4] * 255;
-        data[i + 1] = r * m[5] + g * m[6] + b * m[7] + m[9] * 255;
-        data[i + 2] = r * m[10] + g * m[11] + b * m[12] + m[14] * 255;
-      } else {
+
+      data[i] = r * m[0] + g * m[1] + b * m[2] + m[4] * 255;
+      data[i + 1] = r * m[5] + g * m[6] + b * m[7] + m[9] * 255;
+      data[i + 2] = r * m[10] + g * m[11] + b * m[12] + m[14] * 255;
+      if (!colorsOnly) {
         const a = data[i + 3];
-        data[i] = r * m[0] + g * m[1] + b * m[2] + a * m[3] + m[4] * 255;
-        data[i + 1] = r * m[5] + g * m[6] + b * m[7] + a * m[8] + m[9] * 255;
-        data[i + 2] =
-          r * m[10] + g * m[11] + b * m[12] + a * m[13] + m[14] * 255;
+        data[i] += a * m[3];
+        data[i + 1] += a * m[8];
+        data[i + 2] += a * m[13];
         data[i + 3] =
           r * m[15] + g * m[16] + b * m[17] + a * m[18] + m[19] * 255;
       }
@@ -131,7 +131,7 @@ export class ColorMatrix<
     gl.uniform4fv(uniformLocations.uConstants, constants);
   }
 
-  toObject() {
+  toObject(): { type: Name } & SerializedProps {
     return {
       ...super.toObject(),
       matrix: [...this.matrix] as TMatColorMatrix,