+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [next]
 
+- feat(): Multi select target on events [#10365](https://github.com/fabricjs/fabric.js/pull/10365)
+
 ## [6.5.3]
 
 - fix(ColorMatrix): Restore correct alpha for JS colorMatrix filter [#10313](https://github.com/fabricjs/fabric.js/pull/10313)