+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [next]
 
+- fix(): Reorder caching conditions [#10366](https://github.com/fabricjs/fabric.js/pull/10366)
+
 ## [6.5.3]
 
 - fix(ColorMatrix): Restore correct alpha for JS colorMatrix filter [#10313](https://github.com/fabricjs/fabric.js/pull/10313)