+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [next]
 
-- fix(): Reorder caching conditions [#10366](https://github.com/fabricjs/fabric.js/pull/10366)
+- docs() perf(): Reorder caching conditions for most common scenario and docs fixes. [#10366](https://github.com/fabricjs/fabric.js/pull/10366)
 
 ## [6.5.3]